Refactor: imagefile 패키지에서 file 패키지로 파일 이동
imagefile 패키지는 삭제 예정이므로 사용 중인 파일들을 file 패키지로 이동시킴

resolves: #127
